better compile logic, moved to compile.py  fixes #21
